fix more cases

+++ runtime/include/chpl-comm-compiler-macros.h
@@ -295,7 +295,7 @@ chpl_bool chpl_is_local(chpl_localeID_t loc)
 // and then halt the current task.  (The exact behavior is dictated by
 // chpl_error()).
 static inline
-void chpl_check_local(c_nodeid_t node, int32_t ln, int32_t file, const char* error)
+void chpl_check_local(c_nodeid_t node, int64_t ln, int32_t file, const char* error)
 {
   if (! chpl_is_node_local(node))
     chpl_error(error, ln, file);
